+++ src/main/kotlin/io/openfuture/chain/domain/rpc/transaction/TransferTransactionRequest.kt
@@ -5,7 +5,7 @@ class TransferTransactionRequest : TransactionRequest() {
     override fun getBytes(): ByteArray {
         val builder = StringBuilder()
         builder.append(amount)
-        builder.append(recipientKey)
+        builder.append(fee)
         builder.append(recipientAddress)
         builder.append(senderKey)
         builder.append(senderAddress)