+++ src/main/kotlin/io/openfuture/chain/domain/rpc/transaction/VoteTransactionRequest.kt
@@ -12,6 +12,7 @@ class VoteTransactionRequest(
     override fun getBytes(): ByteArray {
         val builder = StringBuilder()
         builder.append(amount)
+        builder.append(fee)
         builder.append(recipientAddress)
         builder.append(senderKey)
         builder.append(senderAddress)