+++ src/main/kotlin/io/openfuture/chain/domain/transaction/BaseTransactionDto.kt
@@ -3,9 +3,10 @@ package io.openfuture.chain.domain.transaction
 abstract class BaseTransactionDto(
     val timestamp: Long,
     val amount: Double,
+    val fee: Double,
     val recipientAddress: String,
     val senderKey: String,
     val senderAddress: String,
-    val senderSignature: String,
-    val hash: String
+    val hash: String,
+    val senderSignature: String
 )