+++ src/main/kotlin/io/openfuture/chain/entity/Delegate.kt
@@ -27,4 +27,8 @@ class Delegate(
         )
     }
 
+    fun getAddress(): String {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
 }