+++ src/main/kotlin/io/openfuture/chain/entity/transaction/BaseTransaction.kt
@@ -1,10 +1,10 @@
 package io.openfuture.chain.entity.transaction
 
 import com.fasterxml.jackson.annotation.JsonIgnore
+import io.openfuture.chain.crypto.signature.SignatureManager
 import io.openfuture.chain.crypto.util.HashUtils
-import io.openfuture.chain.entity.base.BaseModel
-import io.openfuture.chain.entity.Block
 import io.openfuture.chain.entity.MainBlock
+import io.openfuture.chain.entity.base.BaseModel
 import javax.persistence.*
 
 @Entity
@@ -18,6 +18,9 @@ abstract class BaseTransaction(
     @Column(name = "amount", nullable = false)
     var amount: Double,
 
+    @Column(name = "fee", nullable = false)
+    var fee: Double,
+
     @Column(name = "recipient_address", nullable = false)
     var recipientAddress: String,
 
@@ -27,11 +30,11 @@ abstract class BaseTransaction(
     @Column(name = "sender_address", nullable = false)
     var senderAddress: String,
 
-    @Column(name = "sender_signature", nullable = false)
-    var senderSignature: String,
-
     @Column(name = "hash", nullable = false)
     var hash: String,
+
+    @Column(name = "sender_signature", nullable = false)
+    var senderSignature: String? = null,
 
     @JsonIgnore
     @ManyToOne
@@ -40,15 +43,10 @@ abstract class BaseTransaction(
 
 ) : BaseModel() {
 
-
-
-    fun getBytes(): ByteArray = StringBuilder()
-        .append(timestamp)
-        .append(amount)
-        .append(recipientAddress)
-        .append(senderKey)
-        .append(senderAddress)
-        .append(senderSignature)
-        .toString().toByteArray()
+    @Suppress("UNCHECKED_CAST")
+    fun <T : BaseTransaction> sign(privateKey: ByteArray): T {
+        this.senderSignature = SignatureManager.sign(HashUtils.fromHexString(hash), privateKey)
+        return this as T
+    }
 
 }