+++ src/main/resources/db/migration/V1_5__table_transactions.sql
@@ -1,16 +1,21 @@
 CREATE TABLE transactions (
   id                INTEGER PRIMARY KEY,
-  timestamp         BIGINT NOT NULL,
-  amount            DOUBLE NOT NULL,
-  recipient_address VARCHAR NOT NULL,
-  sender_key        VARCHAR NOT NULL,
-  sender_address    VARCHAR NOT NULL,
-  sender_signature  VARCHAR NOT NULL,
+  timestamp         BIGINT         NOT NULL,
+  amount            DOUBLE         NOT NULL,
+  fee               DOUBLE         NOT NULL,
+  recipient_address VARCHAR        NOT NULL,
+  sender_key        VARCHAR        NOT NULL,
+  sender_address    VARCHAR        NOT NULL,
+  sender_signature  VARCHAR        NOT NULL,
   hash              VARCHAR UNIQUE NOT NULL,
   block_id          INTEGER REFERENCES main_blocks
 );
 --
 CREATE TABLE transfer_transactions (
+  id INTEGER PRIMARY KEY REFERENCES transactions
+);
+--
+CREATE TABLE coinbase_transactions (
   id INTEGER PRIMARY KEY REFERENCES transactions
 );
 --