+++ src/test/kotlin/io/openfuture/chain/block/validation/BlockValidationProviderTests.kt
@@ -37,26 +37,28 @@ class BlockValidationProviderTests : ServiceTests() {
                 VoteTransaction(
                     1500000000L,
                     1000.0,
+                    10.0,
                     "recipient_address",
                     "sender_key",
                     "sender_address",
-                    "sender_signature",
-                    "hash",
                     1,
                     "delegate_host",
-                    9999
+                    9999,
+                    "hash",
+                    "sender_signature"
                 ),
                 VoteTransaction(
                     1500000001L,
                     1002.0,
+                    10.0,
                     "recipient_address2",
                     "sender_key2",
                     "sender_address2",
-                    "sender_signature2",
-                    "hash2",
                     2,
                     "delegate_host2",
-                    11999
+                    11999,
+                    "hash2",
+                    "sender_signature2"
                 )
             )
         )
@@ -88,26 +90,28 @@ class BlockValidationProviderTests : ServiceTests() {
                 VoteTransaction(
                     1500000000L,
                     1000.0,
+                    10.0,
                     "recipient_address",
                     "sender_key",
                     "sender_address",
-                    "sender_signature",
-                    "hash",
                     1,
                     "delegate_host",
-                    9999
+                    9999,
+                    "hash",
+                    "sender_signature"
                 ),
                 VoteTransaction(
                     1500000001L,
                     1002.0,
+                    10.0,
                     "recipient_address2",
                     "sender_key2",
                     "sender_address2",
-                    "sender_signature2",
-                    "hash2",
                     2,
                     "delegate_host2",
-                    11999
+                    11999,
+                    "hash2",
+                    "sender_signature2"
                 )
             )
         )
@@ -131,26 +135,28 @@ class BlockValidationProviderTests : ServiceTests() {
                 VoteTransaction(
                     1500000000L,
                     1000.0,
+                    10.0,
                     "recipient_address",
                     "sender_key",
                     "sender_address",
-                    "sender_signature",
-                    "hash",
                     1,
                     "delegate_host",
-                    9999
+                    9999,
+                    "hash",
+                    "sender_signature"
                 ),
                 VoteTransaction(
                     1500000001L,
                     1002.0,
+                    10.0,
                     "recipient_address2",
                     "sender_key2",
                     "sender_address2",
-                    "sender_signature2",
-                    "hash2",
                     2,
                     "delegate_host2",
-                    11999
+                    11999,
+                    "hash2",
+                    "sender_signature2"
                 )
             )
         )