+++ src/test/kotlin/io/openfuture/chain/block/validation/MainBlockValidatorTests.kt
@@ -2,18 +2,33 @@ package io.openfuture.chain.block.validation
 
 import io.openfuture.chain.config.ServiceTests
 import io.openfuture.chain.entity.MainBlock
+import io.openfuture.chain.entity.transaction.CoinBaseTransaction
 import io.openfuture.chain.entity.transaction.VoteTransaction
-import org.assertj.core.api.Assertions
+import io.openfuture.chain.property.ConsensusProperties
+import org.assertj.core.api.Java6Assertions.assertThat
 import org.junit.Before
 import org.junit.Test
+import org.mockito.BDDMockito.given
+import org.mockito.Mock
 
 class MainBlockValidatorTests : ServiceTests() {
 
+    @Mock private lateinit var properties: ConsensusProperties
+
     private lateinit var mainBlockValidator: MainBlockValidator
+
+    companion object {
+        private const val GENESIS_ADDRESS = "0x00000"
+        private const val REWARD_BLOCK = 10.0
+    }
+
 
     @Before
     fun setUp() {
-        mainBlockValidator = MainBlockValidator()
+        mainBlockValidator = MainBlockValidator(properties)
+
+        given(properties.genesisAddress).willReturn(GENESIS_ADDRESS)
+        given(properties.rewardBlock).willReturn(REWARD_BLOCK)
     }
 
     @Test
@@ -25,36 +40,36 @@ class MainBlockValidatorTests : ServiceTests() {
             "0e09773036394004cb8c340e639a89d7a18e924e8a3d048b49864aeb017e07a0",
             1512345678L,
             mutableListOf(
-                VoteTransaction(
+                CoinBaseTransaction(
                     1500000000L,
-                    1000.0,
+                    10.0 + REWARD_BLOCK,
+                    0.0,
                     "recipient_address",
                     "sender_key",
-                    "sender_address",
-                    "sender_signature",
+                    GENESIS_ADDRESS,
                     "hash",
-                    1,
-                    "delegate_host",
-                    9999
+                    "sender_signature"
                 ),
                 VoteTransaction(
                     1500000001L,
                     1002.0,
+                    10.0,
                     "recipient_address2",
                     "sender_key2",
                     "sender_address2",
-                    "sender_signature2",
-                    "hash2",
                     2,
                     "delegate_host2",
-                    11999
+                    11999,
+                    "hash2",
+                    "sender_signature2"
                 )
             )
         )
 
         val isBlockValid = mainBlockValidator.isValid(block)
 
-        Assertions.assertThat(isBlockValid).isTrue()
+
+        assertThat(isBlockValid).isTrue()
     }
 
     @Test
@@ -69,32 +84,35 @@ class MainBlockValidatorTests : ServiceTests() {
                 VoteTransaction(
                     1500000000L,
                     1000.0,
+                    10.0,
                     "recipient_address",
                     "sender_key",
                     "sender_address",
-                    "sender_signature",
-                    "hash",
                     1,
                     "delegate_host",
-                    9999
+                    9999,
+                    "hash",
+                    "sender_signature"
                 ),
                 VoteTransaction(
                     1500000001L,
                     1002.0,
+                    10.0,
                     "recipient_address2",
                     "sender_key2",
                     "sender_address2",
-                    "sender_signature2",
-                    "hash2",
                     2,
                     "delegate_host2",
-                    11999
+                    11999,
+                    "hash2",
+                    "sender_signature2"
                 )
             )
         )
 
         val isBlockValid = mainBlockValidator.isValid(block)
 
-        Assertions.assertThat(isBlockValid).isFalse()
+        assertThat(isBlockValid).isFalse()
     }
+
 }