+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Fees to transactions
+- New type of transaction: CoinBaseTransaction
+- Genesis address to properties
+- Sign transactions
+- Create and save CoinBaseTransaction
+- Validation of CoinBaseTransaction
+- Constants fees and reward for transactions/block
 
 ## [0.3.0] - 2018-07-13
 ### Added